[2/2][VPlan] Provide an option in TTI to use strict target ISA for matching variants.
This patch updates TTI to provide an opt-in mode where strict
ISA-based matching is done when evaluating available vector-variants.
This ensures that vectorizer does not emit a vector call that
potentially may not be compliant with the ABI of callee. This is
temporary workaround mechanism until a proper fix is developed in
backend to legalize and handle such scenarios.

Change-Id: Iecc9276be99bb8d452c69ba8f6020815c1db0417
JIRA: CMPLRLLVM-27551
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327291
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>