[CMPLRLLVM-28336] Include outer loop preheader for convolution loop nest in HIRRegionIdentification
This change includes the preheader of convolution outer loops in formed HIR
regions, which enables visibility of base pointer loads hoisted to that
preheader in cases where the outermost loop can't be included in the HIR
region due to the presence of atomics. It also fixes a related bug where
handling non-loop block region live outs would result in segfaults if
the region exit block was a loop latch.

Change-Id: Ib9c75d900253dbc134c9858e3acd598155828ceb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327255
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>