[DTrans - OpaquePtr] DeleteFieldOP pass to create types and process GEP instructions
Create the new structure type and update the index on GEP instructions that will
use the new type. Delete the GEP for the case where the field is being removed.

Much of this code is taken from the existing DeleteField pass,
but with changes needed to work with the new DTrans opaque
pointer infrastructure. The intent is that the original
delete field pass will be completely replaced by this at some
point, so there was not an attempt to break that into smaller
functions to share code between the new and old implementations.

Test cases are named to match up with the original delete field
test cases to make it easy to identify which tests have been
converted to the new pass.

Change-Id: Ie07f352e50a3b1ff623632af53c84ca7c1991cbe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327574
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>