[DTrans - OpaquePtr] DeleteFieldOP pass to process byte-flattened GEP instructions
Updates the offsets for byte-flattened GEPs for structures that are changing.

Much of this code is taken from the existing DeleteField pass,
but with changes needed to work with the new DTrans opaque
pointer infrastructure. The intent is that the original
delete field pass will be completely replaced by this at some
point, so there was not an attempt to break that into smaller
functions to share code between the new and old implementations.

Test cases are named to match up with the original delete field
test cases to make it easy to identify which tests have been
converted to the new pass.

Change-Id: I5fb3259eb2106eb1d6689ac40eb67cea80e39711
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327575
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>