[DTrans - OpaquePtr] DeleteFieldOP pass to update structure size values used for function calls
Much of this code is taken from the existing DeleteField pass,
but with changes needed to work with the new DTrans opaque
pointer infrastructure. The intent is that the original
delete field pass will be completely replaced by this at some
point, so there was not an attempt to break that into smaller
functions to share code between the new and old implementations.

Test cases are named to match up with the original delete field
test cases to make it easy to identify which tests have been
converted to the new pass.

Change-Id: I9c76cfbf89ae17079a6299ef4b0f32067c53c639
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327576
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>