[SYCL][FPGA] Fix num_simd_work_items and reqd_work_group_size argument check  (#3728)
The arguments to reqd_work_group_size are ordered based on which index
increments the fastest. In OpenCL, the first argument is the index that
increments the fastest, and in SYCL, the last argument is the index that
increments the fastest.

In OpenCL, all three arguments are required.

In SYCL, the attribute accepts either one, two, or three arguments; in each
form, the last (or only) argument is the index that increments fastest.
The number of arguments passed to the attribute must match the dimensionality
of the kernel the attribute is applied to.

If the reqd_work_group_size attribute is specified on a declaration along
with num_simd_work_items, the required work group size specified
by num_simd_work_items attribute must evenly divide the index that
increments fastest in the reqd_work_group_size attribute.

This patches fixes the problem on https://github.com/intel/llvm/pull/3275 where we checked the wrong argument for SYCL and OpenCL attributes.

Fixes https://jira.devtools.intel.com/browse/CMPLRLLVM-28332

Change-Id: I74d9fb9fb4c4f8cc57c8a94ac8471fe3dd4f62e8
Signed-off-by: Soumi Manna <soumi.manna@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327622
Reviewed-by: Aaron Ballman <aaron.ballman@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>