[SYCL][FPGA] Fix num_simd_work_items and reqd_work_group_size argument check  (#3728)
The arguments to reqd_work_group_size are ordered based on which index
increments the fastest. In OpenCL, the first argument is the index that
increments the fastest, and in SYCL, the last argument is the index that
increments the fastest.

In OpenCL, all three arguments are required.

In SYCL, the attribute accepts either one, two, or three arguments; in each
form, the last (or only) argument is the index that increments fastest.
The number of arguments passed to the attribute must match the dimensionality
of the kernel the attribute is applied to.

If the reqd_work_group_size attribute is specified on a declaration along
with num_simd_work_items, the required work group size specified
by num_simd_work_items attribute must evenly divide the index that
increments fastest in the reqd_work_group_size attribute.

This patches fixes the problem on https://github.com/intel/llvm/pull/3275 where we checked the wrong argument for SYCL and OpenCL attributes.

Fixes https://jira.devtools.intel.com/browse/CMPLRLLVM-28332

Change-Id: Idc8f9e773a42e094d7c9d85d42d553c3403c5766
Signed-off-by: Soumi Manna <soumi.manna@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327620
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>