Fix incorrect merge.
Opt-report infracture was added in xmain for openMP and in intel/llvm
for SYCL. When the intel/llvm patch was cherry-picked (57c7a5a8761)
(https://git-amr-2.devtools.intel.com/gerrit/#/c/323602/),
the 2 implementations were combined since they shared common infrastructure.
However, when the SYCL patch came in later via the nightly SYCL merge, the
git merge or conflict resoltion done kept incorrect remnants of the SYCL
implementation (which were removed during cherry-pick), in xmain.

This patch fixes the incorrect merge to keep what was done in original
cherry-pick. It also adds INTEL_CUSTOMIZATION markers which were missed
during the cherry-pick.

In order to avoid future conflicts of this nature, we should refactor
xmain or upstream to make it as similar as possible. This will be done
in a followup patch.

Change-Id: I7a01b45d6be84610e3586f3b59e1f284e531c464
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327637
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>
Reviewed-by: Rao, Premanand M <premanand.m.rao@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>