[OptReport][AsmPrinter] Prototype for Protobuf-based binary opt-report feature in Xmain.
Xmain backend currently uses the legacy bitset based approach
to encode text opt-report remarks into binary opt-report format.
This approach has some known drawbacks like limitation on number
of representable remarks/properties, and explicit version tracking
for backward compatibility between client and compiler. In this
patch we propose a new Protobuf-based solution to encode compiler's
text opt-report as a serialized and compressed binary message which
can be embedded in generated object file. A central Protobuf schema
is designed to define the format/rules for serializing compiler's
opt-report; this same schema is expected to be used by clients to
deserialize the binary opt-report. We also simplify layout of the
".debug_opt_report" section and introduce the concept of "anchor
table" to map a specific loop opt-report to it's appropriate source
location in the object file.

JIRA : CMPLRLLVM-23604

Change-Id: Ie48a5500dfdf45fe2e82040c404203105e7842dc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321838
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>