[SLP] Fix getGatherCost() crash after rGab2c499d3a2e: [SLP] Add insertelement instructions to vectorizable tree
The change is deemed a work around fix. InsertElement instructions are
considered vectorization tree seeds, thus only appear as a tree root node.
If the node is marked as NeedToGather for whatever reason it basically
means the entire tree is not vectorizable and isTreeTinyAndNotFullyVectorizable()
supposed to return true for it. Due to current customizations cost
calculation happens regardless of the routine answer and fails on such entries.
This patch just fixes the crash via correctly picking type of scalars.
Calculated cost will not be used anyway.
Fixes CMPLRLLVM-28448, CMPLRLLVM-28462.

Change-Id: I5b16205a41214b6166f61f7ec2f7f693e4f93bce
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327736
Reviewed-by: Alexey Bataev <alexey.bataev@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>