[NFC][loopopt] Preparation to make eliminateRedundantGotos a utility that can be called during HIR vector code generation
HIR vector code generation is planning to use code implemented as part of
HIR construction to eliminate redundant gotos and labels. This is the first
step in that direction. The changes here make the goto elimination independent
of BasicBlock information so that this can be used from HIR VPlan vectorizer.

Change-Id: I6340a210ed506dec128f64e10cb9a239786eaf23
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327785
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>