[Paropt] Capture directives when they are non-first BB instructions while adding privates.
Currently the utility addPrivateToEnclosingRegion handles cases
when the directive intrinsic is found as the first instruction
in the BB being visited. This patch relaxes this restriction by
upating the utility to iterate over instructions in the BB and
identify any instruction that represents an OMP directive. This
ensures correctness of middle-end transforms which rely on
privatization after within loop allocas are hoisted out.

JIRA : CMPLRLLVM-28500

Change-Id: I8812d2541cd1ae7b5c862922edb4e0b0d6a5c332
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/328099
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>