[DTrans] [CMPLRLLVM-28202] DTrans delete field pass to process load/store alignment
The alignment argument can be placed on a load/store instruction by the front-end
or another pass that runs prior to DTrans. When this is done for structure fields
due to information known about the alignment of a pointer to the structure and
the layout of structure members, these alignment values may no longer hold true
following the deletion of field members. This caused a problem that an instruction
was used which requires a memory address to be aligned on a 64-byte boundary, but
this could no longer be guaranteed following the removal of fields. It is not possible
to simply prevent the delete field pass from processing structures that have had
load/store operation marked with alignments that are higher than the default value
for the type, because that would impact cases that require field deletion This change
updates the alignment values to their native values for the load/store instructions
impacted by the transformation, and then lets a downstream invocation of the
instcombine pass to again increase above the default alignment for cases that are
provable based on the updated structure layout.

Change-Id: I6d84c1e7aef47f6fa93010547201f3ce1c9b20e7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/328278
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>