[DTrans] Alternative solution to CMPLRLLVM-26963 to prevent delete field crash
The change https://git-amr-2.devtools.intel.com/gerrit/#/c/318022/ had an
unexpected side effect on the delete field optimization pass. Specifically,
that change causes the delete field candidate selection to evaluate dead types,
such as:
  LLVM Type: %"type 0x2343f30"
    Can delete field: %"type 0x2343f30" @ 0

This is a type that has been removed from the IR, but still existed in memory
because there were pointers to the object. This occurs because there was some
basename version of the type that can be retrieved by StructType::getTypeByName,
but which is not discoverable by walking the IR. For example:
  %struct.basename = {  %"type 0x2343f30"* }
These dead types exist because when the IRMover loads bitcode files, a type
may come from a bitcode file that will be discarded because another bitcode
file already had the type.

When these dead types are made visible to the delete field pass, it cause the
delete field pass to process the IR even when there are no types in use which
can have fields deleted because it causes the pass to produce more unused types,
such as:
 %__DFT_.3548 = type { %__DFDT_.3561, %"type 0x23559b0", %"type 0x23559e0", %"type 0x2355a10" }

This modifies the implementation of the referenced changeset to fix the DTrans
crash by recognizing that the original enclosing type map produced from a set
of all types in the symbol table may contain types that were not reachable during
the safety analysis. In this case, those types do not need to be evaluated for
safety.

No new LIT tests are added with this change because the observable behavior is
no different than ensuring that the test case deletefield47.ll from
https://git-amr-2.devtools.intel.com/gerrit/#/c/318022/.
does not crash the pass.

Change-Id: I468ef7325a1ea1f350c9362438816b9f9effb3ca
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/328277
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>