CMPLRLLVM-28636: adjust LIT checking for Intel-only subscripts RUN
The final section of checks in this test verify function attributes.
The convert-to-subscripts output has an additional function attribute
because of its use of the subscripting intrinsic. Consequently, the
CHECK-NEXT before the subscript intrinsic's attribute line must be a
CHECK in order to allow the subscript RUN to ignore the sebsequent line.

Change-Id: Ibbc1f3df66d60e8a4adfeaf3b3cf02129acfa6a7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/328455
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>