Cherry pick https://github.com/intel/llvm/pull/3797 to xmain-rel
[SYCL]Fixes a performance issue due to
https://github.com/intel/llvm/pull/3612

This change does two things to fix the performance issue.
The first is in the level zero plugin. Here the change is to only
close and submit the batch if the event being queried is one that
will be signalled by one of the commands in the batch.

The second change is in the sycl run-time itself. This change
prevents the event cleanup code from querying every single event
that is outstanding in the system. This is necessary to prevent
the most recent events (which are likely to be in the plug-ins batch)
from being queried.

Both changes are required to fix the performance issue.

The performance regression was reported as CMPLRLLVM-28493.

Change-Id: Ia7ff8ede592b078546201c2cb6bda9c75ea5858e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/328546
Tested-by: Kevin B. Smith <kevin.b.smith@intel.com>
Reviewed-by: Kevin B. Smith <kevin.b.smith@intel.com>