[Paropt] Emit remark if map clause for a struct variable can be changed to a firstprivate
This patch extends map clause analysis to structure types in addition to scalars. Compiler
will print remark if map clause for a structure variable can be changed to a firstprivate.
This should be legal to do if variable address is not escaped and it is not modified inside
the target region.

Change-Id: Ie978cd31857a44fb82e3b30443b7f296d6c62af8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/328886
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>