[DTrans] [CMPLRLLVM-28673] Fix delete field crash when global variables have initializers
When the delete field pass was extended to allow structures that are nested
types or contain nested types, the code that handles processing global variables
was not updated to handle generating the replacement of non-zeroinitializers.
This resulted in a compiler crash, unless fields were deleted from both the nested
type, and the type containing it.

Change-Id: Ica9239c6778e02957cd167fb0291b33efe5260d2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/329081
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>