[DTrans - OpaquePtr] DeleteFieldOP to handle global variables.
Replace global variables, update initializers and update GEPOperators
of global variables that are changed.

Much of this code is taken from the existing DeleteField pass,
but with changes needed to work with the new DTrans opaque
pointer infrastructure. The intent is that the original
delete field pass will be completely replaced by this at some
point, so there was not an attempt to break that into smaller
functions to share code between the new and old implementations.

Test cases are named to match up with the original delete field
test cases to make it easy to identify which tests have been
converted to the new pass.

Change-Id: Ia1a40d867ceef2c50c47820ea6b7184fe2945a71
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/329086
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>