[LocalBuffers] Attach __local debug info to LocalMem to avoid memcpys
Why:
- OpenCL __local variables are represented as GlobalVariables in llvm IR
to be visible at work group scope (this is done by FE).
- OCL CPU BE actually maintains the __local vars in the LocalMem buffer
(either represented by a TLS global or an implicit arg passed to the
kernel).
- The debug info of a __local var, which is generated by FE, was
attached on the GlobalVariable. But the real value of __local resides in
the LocalMem buffer. The LocalBuffers used explicit memcpys from
LocalMem to the GlobalVariable in each barrier loop latch, to guarantee
that a debugger could read the real value of __local vars.
- The memcpys degraded the '-g' program performance significantly.

How:
- These memcpys can be removed, as long as we transfer the __local debug
info from GlobalVariable to LocalMem buffer (where the real value
resides) to keep the debugging functionality.
- The offset of each __local var in the LocalMem buffer is known at
compile time. So we can easily attach the debug info to LocalMem with
corresponing buffer, encoded in DWARF expressions (e.g. deref, plus).
Such that we improve the '-g' performance and keep the debuggability as
well.

What:
- Original debug info on GlobalVariable is transferred to the function
level, as a debug local variable (AutoVariable) created by compiler.
- The new debug info is attached onto the LocalMem buffer pointer, with
a (deref + offset) DWARF expression, so that the debugger is responsible
for finding __local var's real value during debugging.
- Original GlobalVariable is also safely removed, after replacing all
its uses.
- The `DebugCopy()` call instruction generating in Barrier Pass is also
removed. It was used as an anchor to insert memcpys.
- Refactor LocalBuffers Pass to improve readability.
- Remove LocalBuffers wrapper classes. It prevented "-print-after/before"
and "-debug-only" from working for LocalBuffers pass.
- Extend debugger_test_type/test_barriers.py to assure the debuggability
on __local vars.
- Re-enable debugger_test_type/test_variable_setting.py as users can set
values of __local vars in debugger now.

Change-Id: I9f35d9b1afcebe4555199156a9360bbe707e57eb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/329369
Tested-by: Yilong Guo <yilong.guo@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>