CMPLRLLVM-28733: Revert bad community change set
The latest pulldown picked up a change to Argument Promotion which leads
almost any program with OMP pragmas to die inside Argument Promotion. The
change has already been reverted in llorg, but there is a request to fix
it now in xmain because of the massive number of regressions.

This a partial reverting of the communit revert:
  https://reviews.llvm.org/D100917

The change to intel-sub-to-add-prelto-suppress.ll is needed because
FileCheck is now matching the 'addr' string in the function attributes,
causing the 'CHECK-NOT: add' to fail.

Change-Id: I7550d65775c10d7afcbe9a0293280872c2d37d7e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/329413
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>