[VPlan][CFGMerge] Make cloning of scalar peel instead of scalar remainder.
We do clone original scalar loop when we use it for both peel and remainder.
Cloning of scalar remainder leads to incorrect code generation due to
  - the cloninig is done at code generation
  - code generation for peel is done earlier (as it's earlier in cfg)
  - creating peel, we replace the upper bound of the loop and that causes
    cloned remainder to incorrectly use the same changed upper bound
Cloning the loop for scalar peel resolves that problem because in this
case we change the upper bound of clone, leaving the original one untouched.

Change-Id: Ie21784bd3f5aa1f7bb455e6787ff8d3d5b14777f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/329759
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>