[loopopt] Cleanup redundant HLIfs in HIRCleanup phase of the framework.
Some conditional branches which can be proven to be true or false
based on identical dominating conditions are not cleaned up by ScalarOpt
before HIR because of implementation limitations.

This changesets adds similar logic in HIRCleanup as a workaround.

This is a more generalized version of a previous workaround implemented
in HIRLoopFormation phase which has now been removed.

Both are required to handled incoming IR in 503.bwaves.

ValueTracking's isImpliedByDomCondition() functionality was extended
to use DominatorTree to be able to handle these cases.

The functionality can be moved to SimplifyCFG when it is mature enough
to use DominatorTree analysis.

Change-Id: Ib61b7376bf046f8f0f0fbf6ebde3839e75a92cfe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/329928
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: Chawla, Pankaj <pankaj.chawla@intel.com>