Fix the Support/HomeDirectory unittest on Windows
The Support/HomeDirectory unittest retrieves the "USERPROFILE"
environment variable path and compares it to the path returned by the
LLVM-specific path::home_directory() utility function. This is meant to
verify that path::home_directory() is working correctly. However, if the
environment variable is in posix-style format (as it is when the test is
run in a Unix-shell on Windows) the comparison fails.

We had previously disabled this test in xmain. This change re-enables
the test and updates it to convert the value retrieved from the
environment variable to native style before the comparison.

This fixes CMPLRLLVM-27652.

Change-Id: I6e20ba312180468a71c28526e5f3f3918fec2208
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/330105
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>