[DTrans - OpaquePtr] Initial changes for DTransSafetyAnalyzer field value collection
This is part 1 of the changes for the DTransSafetyAnalyzer to collect values
stored to fields. This change introduces the basics of value collection, by
collecting values that are used for store instructions. It also, marks all
fields of structures that do not pass the safety checks as having an incomplete
value set.

Change-Id: I9f1131f3ec9d843cbad54c54b4c42a90e290b624
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/330129
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>