CMPLRLLVM-28810: Inhibit callback cloning for varargs functions
Callback cloning was not intended for varargs functions. This change set
inhibits it explicitly, and adds a line in the trace to indicate whether
callback cloning was attempted. It also fixes a previous mistake, where
callback cloning was forced on by default. The correct default is that
callback cloning is neither forced on or off by default. I also added an assert
to ensure that it is not both forced on and off at the same time.

Change-Id: I777aa75e8282e990d08dca07e94908de209f1a82
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/330256
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>