[VPlan] Relax restrictions on latch condition during normalized induction analysis.
With some combinations of loop latch condition and latch successors the upper
bound used in the latch condition is not equal to loop iteration count.
E.g.
    ...
    %c = icmp ule ind_incr, N
    br %c, label %loop_body, label %loop_exit

The loop iteration count is equal to that upper bound + 1 in such cases, and
we consider such loops as not having normalized induction.
This change relaxes that restriction by adding special code to process such
loops in cloning for masked mode, inductoin last value calculation, and
loop induction emission.
Peeling is (temporarily) disabled for such loops - we need some special code
to update the upper bound in such loops.

Change-Id: I02704ad9cdcad9c2e639e3030549f66417104dab
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/330649
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>