[NFC][Ordered SIMD] Skip VPlanPragmaOmpOrderedSimdExtract pass in opt-bisect mode.
Currently VPlanPragmaOmpOrderedSimdExtract pass is run
unconditionally invoked from IPO PassManager pipeline.
However in-order to support debuggging with opt-bisect
we should be able to skip the pass if OptPassGate
instructs us to not run the pass. This patch enables
this support by adding a skipModule check in the pass.

Change-Id: I2654cff1a27f263c0cde41ac0f945ae6d763788f
JIRA: CMPLRLLVM-28706
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/330782
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>