[Driver] Improve consistency of pulling in Intel headers for OpenMP
We were pulling in the Intel SYCL headers for OpenMP device compilations
very late in the command line sequence, causing the system ones to be
picked up instead of the Intel ones.  This caused a mismatch of host
and target headers being used when doing a combined SYCL/OpenMP interop
compilation.

Update the header inclusion to always be in the same area for both SYCL
and OpenMP usage paths.

Change-Id: I05a702d52b48aebde10ede179b1c6990a19c73fd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/330881
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>