[NFC] CMPLRLLVM-29032 and CMPLRLLVM-29040: Klocwork fixes 20210606
The change in Intel_DopeVectorAnalysis.cpp is because Klocwork thinks that
GEPO could be nullptr.  In fact, it can't happen there, because we have
already checked that the stride part of a SubscriptInst is fed by a constant
or a GEP. But I am adding the check in case someone else decides to use this
routine in a different context in the future.

The assert added to SOAToAOSEffects.h is valid because computeDepApproximation
has already been called to compute the approximation before getApproximation
is called to get it.

Change-Id: I51517556e627f4260b56ae685e9f4d8208484552
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331049
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>