[NFC] [DTrans] Modify interface to DTransAnnotator functions that create metadata annotations to prepare for opaque pointers
The DTrans annotations that are used to mark DTrans transformation information
for subsequent DTrans passes stores a pointer type in the metadata. Because
opaque pointers will make different pointer types indistinguishable from one
another, a different encoding will be necessary. This modifies the interface
functions used to create the metadata to prepare for this change, but does
not make any functional changes.

Change-Id: I46f5f55f3217cbbce8c5bc0ca6a21baf12691d1e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331254
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>