[NFC] [DTrans] Modify interface for extracting/checking DTrans annotations to prepare for opaque pointer
The DTrans annotations that are used to mark DTrans transformation information
for subsequent DTrans passes stores a pointer type in the metadata. Because
opaque pointers will make different pointer types indistinguishable from one
another, a different encoding will be necessary. This modifies the interface
functions called to extract and check for the existence of the metadata to
prepare for this change, but does not make any functional changes.

Change-Id: I79e02376806e0c8fe6bce0067374b9c442b404aa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331255
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>