[DTrans] Update metadata encoding used by DTrans annotations to prepare for opaque pointers
The DTrans annotations that are used to mark DTrans transformation information
for subsequent DTrans passes stored a pointer type in the metadata. Because
opaque pointers will make different pointer types indistinguishable from one
another, a different encoding is necessary. This modifies for encoding
format to avoid directly storing the pointer types in the metadata.

Change-Id: Ieb5b5bb1bdcc64f481a16a442a78f4ad6664cf28
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331256
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>