[OpenCL] Add memory_scope_all_devices
Apply change to opencl-c-common.h and this fixes CMPLRLLVM-29021.

We are splitting opencl-c.h into opencl-c-20.h, opencl-c-common.h,
opencl-c-platform-12.h, opencl-c-platform-20.h and opencl-c-platform.h in
xmain. This is to decrease memory overhead when the header is compiled into
PCH, because opencl-c.h relied on the _OPENCL_C_VERSION, and declared some of
the built-ins differently for OpenCL 1.2 and 2.0. When the header is compiled
into PCH, only one value of the __OPENCL_C_VERSION_ can be used. To support
both OpenCL C 1.2 and 2.0, it was necessary to compile the header twice with
different values of the _OPENCL_C_VERSION_ macro. As a result, both of the
compiled headers contained a copy of the 'common' built-ins. So we have such
change for opencl-c.h.
And opencl-c.h is splitted into opencl-c-base.h and opencl-c.h in community
for use of fdecalare-opencl-builtins option. We have also same
opencl-c-base.h. When there is change from community, we have to apply change
in both two parts now. This is because opencl-c-base.h is formed by moving
builtin types and constants from opencl-c.h, and there are various type for
all opencl versions in it.

Change-Id: I3793dfb00f70caa8f01c58d80a7a6ad2bc51dbb3
Signed-off-by: Haonan Yang <haonan.yang@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331071
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>