[VPlan][byval] Callee processing in VecClone
Having a vector pointer argument means that a vector of pointers will be
attempted to be constructed. However, if this argument is marked as
byval, then converting pointer to a vector violates byval semantics.

The proposed transformation is this. The original argument:
    %struct.A* byval(%struct.A)
will be turned into
    [4 x %struct.A]* byval([4 x %struct.A])

The array is used since it is not allowed to construct a vector out of
non-primitive types.

ICC compiler disallowed using structs passed by value within OpenMP,
perhaps for performance reasons. ICX does not do this.
Moreover, structs passed by value are often seen in the code generated
from DPCPP. So if we want to support function vectorization within
DPCPP, we have to be able to handle that.

Change-Id: I85e7e9cbd4a7602a8a3920e5eb6e6d07a0c6ec62
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/332859
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>