[OpenCL] Sync opencl headers and loader with upstream
Change removes INTEL_CUSTOMIZATION and aligns OpenCL Headers and Loader
dependencies to the same version upstream (intel/llvm github) is using.

We used to use special ocl-open branch for that purpose that's been
updated by the pulldown tools every time new pulldown candidate is
created. This is not working this way anymore, opencl deps are not in
manifest list and instead of pulldown tool doing the versions uplift
this will be done naturally by getting right hashes in CMakeLists from
the upstream.

Change-Id: I65b4a3335dde04d9eaf304799a6dccc9d92d10bb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331168
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>