[OpenMP] Remove standard CMake output vars override
These settings override CMAKE defaults. These are producing problems for any
dependencies libomptarget may depend on, e.g. libOpenCL.so. It leads to
non-determenistic placement of libOpenCL.so in the build directory.
Basically it is determined by the project that reaches dependency first.
Some internal unit tests and tools expect libOpenCL.so looking up for library
in the standard llvm build directory.
Disabling for now to test more.

This is fix for CMPLRLLVM-28949

Change-Id: I084cd07878019518ef973cf23379bb7478759d8b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331219
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Bae, Hansang <hansang.bae@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>