[OpenCL] Fix subgroup emulation issue
The kernel is failed to compile with undefined function error. This failure
happens when there are two kernels calling the same sub group builtin. One
kernel is vectorized and the builtin is cloned to widened function. But another
is not vectorized, and handled by subgroup emulation. The later will try to
widen the same builtin but have to add postfix since there is already same widen
version of function.

Fixed by skipping widening function if the widened version of function
exists.

Change-Id: I778426d80f2a3044f20e1127fe6305a443822bb1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331249
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: wuwei <wei.wu@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>