[VPlan][CFGMerger] Remove unneeded assertion on absence of remainder.
In case we have a known tc and static peel, we can calculate exactly
whether we need a remainder. The cfg-merger had an incorrect assertion
requiring a remainder always when we have any peel.

Also, added a work around to add scalar remainder if we decided not
having any remainder with unknown trip count.

 Part of the fix for cmplrllvm-29120.

Change-Id: Iffebc9738a571ea0ca2b62ee4a48fad2f35dfc5f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331348
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>