CMPLRLLVM-28726: Add math functions prototype if LTO is enabled
This change set adds a path that inserts the math function
prototypes if the llvm intrinsic form are in the IR and the
compile step is generation an LTO (bitcode) object. The idea
is to insert undefined math symbols in the IR, then the linker
should be able to pull the Intel's math libraries in the first
symbols resolution and generate the executable using fast math.

This patch fixes the multiple regressions after moving libimf
from dynamic linking to static linking:

  * spec_omp2012ref/352.nab   +12.39%
  * cpu2017/644.nab           +13.27%
  * specHPC2021ref/soma.513   +4.92%

The patch can produce new possible gains since we are prioritizing
Intel's math libraries over GNU libraries:

  * cpu2017/538.imagick       +1.80%

Change-Id: I6e083b010f60203f68cc2214faabfdeae318879d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331434
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>