CMPLRLLVM-28297: Dead loops and store elimination (part 1)
This is the first change set to enable dead loops and store
elimination. These changes enables the argument promotion
at the early stages of LTO to clean the IR. This pass
simplifies the IR and enable other optimizations like
constant propagation, sparce constant propagation and dead
arguments elimination.

This change set also includes a fix in tiling for multiversioning.
Due to argument promotion, it can hoist load instructions from
the called function to the caller function. The fix in tiling
is to catch this behavior too.

Change-Id: I700e276bba3cf76ed3004bbb46a70024f6bb1733
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331893
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>