[NFC][VPlan] Update tests in preparation for privates support along HIR path.
We need to update tests for two unavoidable reasons,
1) HIR-path does not support last-privates. So, SOA-analyses give
different result along HIR-path. To make the test-command common for
both IR and HIR path we need to get rid of last-privates.
2) HIR is not constructed for loops with 'volatile' load/store instructions,
this means we cannot reliably do escape-analysis for IR with volatile
loads/store on privates. So, the existing test has to be put in the
different file.

Change-Id: I491f3d3d2f5ff9a763b542754988e6559878cd3a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/331663
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>